refactor(inflight): extract delivery rewind save
