+++ src/services/discord/inflight/save_store.rs
@@ -13,18 +13,21 @@
 
 use super::*;
 
+#[path = "save_store/delivery_rewind.rs"]
+mod delivery_rewind;
 #[path = "save_store/identity_gate.rs"]
 pub(super) mod identity_gate;
 #[path = "save_store/rebind_adoption.rs"]
 mod rebind_adoption;
 
+pub(in crate::services::discord) use self::delivery_rewind::save_inflight_delivery_rewind_if_matches_identity;
 pub(in crate::services::discord) use self::identity_gate::{
     GuardedSaveOutcome, bind_recovery_anchor_if_matches_identity,
     mark_readopted_from_inflight_if_identity_unchanged,
     patch_restart_full_response_if_identity_unchanged,
     persist_leak_recovery_response_offset_if_matches_identity_locked,
     persist_recovery_output_path_if_matches_identity_locked,
-    recovery_anchor_msg_id_if_matches_identity, save_inflight_delivery_rewind_if_matches_identity,
+    recovery_anchor_msg_id_if_matches_identity,
     save_inflight_state_if_identity_matches_allow_output_restamp,
     save_inflight_state_if_identity_unchanged, save_inflight_state_if_matches_identity,
     stamp_claude_e_process_if_matches_identity,
